reorganized repl
