+++ website/src/components/Udels/UdelsEditor.jsx
@@ -1,4 +1,3 @@
-import { ReplContext } from '@src/repl/util.mjs';
 
 import Loader from '@src/repl/components/Loader';
 import { Panel } from '@src/repl/components/panel/Panel';
@@ -8,20 +7,15 @@ import UserFacingErrorMessage from '@src/repl/components/UserFacingErrorMessage'
 
 // type Props = {
 //  context: replcontext,
-//  containerRef:  React.MutableRefObject<HTMLElement | null>,
-//  editorRef:  React.MutableRefObject<HTMLElement | null>,
-//  error: Error
-//  init: () => void
 // }
 
 export default function UdelsEditor(Props) {
-  const { context, containerRef, editorRef, error, init } = Props;
-  const { pending, started, handleTogglePlay } = context;
+  const {context} = Props;
+  const { containerRef, editorRef, error, init, pending, started, handleTogglePlay } = context;
+
   return (
-    <ReplContext.Provider value={context}>
       <div className={'h-full flex w-full flex-col relative'}>
         <Loader active={pending} />
-        {/* <Header context={context} /> */}
         <BigPlayButton started={started} handleTogglePlay={handleTogglePlay} />
         <div className="grow flex relative overflow-hidden">
           <Code containerRef={containerRef} editorRef={editorRef} init={init} />
@@ -29,6 +23,5 @@ export default function UdelsEditor(Props) {
         <UserFacingErrorMessage error={error} />
         <Panel context={context} />
       </div>
-    </ReplContext.Provider>
   );
 }
